requirements.txt
discord.py
yt-dlp
spotipy
lyricsgenius
aiohttp
pynacl